fix(priority): use unguess table for priorities tables + rename table in unguess_table
