+++ src/__mocks__/database/campaign_read_status.ts
@@ -1,4 +1,4 @@
-import Table from "./table";
+import Table from "./unguess_table";
 
 interface CampaignReadStatusParams {
   id?: number;
