+++ src/__mocks__/database/customer_unique_bug_read.ts
@@ -1,4 +1,4 @@
-import Table from "./table";
+import Table from "./unguess_table";
 
 type CustomerUniqueBugsReadParams = {
   wp_user_id?: number;
