+++ src/__mocks__/database/template_categories.ts
@@ -1,4 +1,4 @@
-import Table from "./table";
+import Table from "./unguess_table";
 
 type TemplateCategoryParams = {
   id?: number;
